… image_queries/Day5/query for animal ID 4 more less than 1s.png
… image_queries/Day5/query for animal ID 4 more than 1s.png
… image_queries/Day5/query for owners email less than 1s.png
… image_queries/Day5/query for owners email more than 1s.png