# Pull Request Review

## In this milestone we have: 

- Added a new column to the owner's table with the name 'email'.
- Insert into visits and owners table new data.
- Run EXPLAIN ANALYZE on three selected queries.

# BEFORE:

## Query 1
<img width="577" alt="query for animal ID 4 more than 1s" src="https://github.com/edgardsoza/vet-clinic-database/assets/105621110/9c084f84-e567-427d-810d-70fe348ef2d6">

## Query 2
![2nd query more than 1](https://github.com/edgardsoza/vet-clinic-database/assets/105621110/1b147ec7-3c73-48fd-8254-f20c1eead512)

## Query 3
<img width="514" alt="query for owners email more than 1s" src="https://github.com/edgardsoza/vet-clinic-database/assets/105621110/27558a42-b0a1-4bea-9b2e-2af91f9692b6">

# AFTER:

## Query 1 
<img width="674" alt="query for animal ID 4 more less than 1s" src="https://github.com/edgardsoza/vet-clinic-database/assets/105621110/576417d2-5a78-47ce-8bc7-30b23452c352">

## Query 2
![2nd query less than 1](https://github.com/edgardsoza/vet-clinic-database/assets/105621110/83ada9d8-eb36-4fe7-b327-8eef0b99676c)

## Query 3
<img width="501" alt="query for owners email less than 1s" src="https://github.com/edgardsoza/vet-clinic-database/assets/105621110/fdbe23bb-9ed5-4fc0-b928-05820cc0f76f">


Please let us know if there´s anything else we should add to our project.

Thank you

Regards,
The Team.